+++ src/main/java/com/jnape/palatable/lambda/semigroup/builtin/LeftAll.java
@@ -37,7 +37,7 @@ public final class LeftAll<L, R> implements SemigroupFactory<Semigroup<L>, Eithe
 
     @Override
     public Semigroup<Either<L, R>> apply(Semigroup<L> lSemigroup) {
-        return (x, y) -> x.flatMap(xL -> y.flatMap(yL -> left(lSemigroup.apply(xL, yL)), Either::right), Either::right);
+        return (x, y) -> x.match(xL -> y.match(yL -> left(lSemigroup.apply(xL, yL)), Either::right), Either::right);
     }
 
     @SuppressWarnings("unchecked")